% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-AM.R
\name{ascend}
\alias{ascend}
\title{Perform A Single Gradient Ascent}
\usage{
ascend(model, input, loss, gradient, lr = 1)
}
\arguments{
\item{model}{A keras model.}

\item{input}{The input to perturb.}

\item{loss}{The output from \code{\link{get_layer_loss}}.}

\item{gradient}{The output from \code{\link{get_layer_gradient}}.}

\item{lr}{The learning rate used to update input.}
}
\value{
The perturbed input.
}
\description{
Perform A Single Gradient Ascent
}
\examples{
library(keras)
library(caress)
data(iris)
x <- as.matrix(iris[,1:4])
y <- to_categorical(as.numeric(iris[,5])-1)
input <- from_input(x, name = "input")
output <- input \%>\%
  layer_dense(units = 2, activation = "linear", name = "middle") \%>\%
  to_output(y, name = "output")
model <- prepare(input, output)
build(model, x, y, epochs = 10, batch_size = 8)
loss <- get_layer_loss(model, "middle")
gradient <- get_layer_gradient(model, loss)
x <- matrix(runif(4), 1, 4)
for(i in 1:20){
  x <- ascend(model, x, loss, gradient)
}
}
