% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-build.R
\name{build}
\alias{build}
\title{Train a Keras Model}
\usage{
build(model, x_train, y_train, lr = 0.001, epochs = 30,
  batch_size = 128, patience_early_stopping = 5,
  validation_split = 0.2)
}
\arguments{
\item{model}{A keras model.}

\item{x_train, y_train}{A numeric or list of numerics.
Arguments passed to \code{keras::fit}.}

\item{lr}{Arguments passed to \code{keras::compile}.}

\item{epochs, batch_size, validation_split}{Arguments passed to \code{keras::fit}.}

\item{patience_early_stopping}{Argument passed to
\code{keras::callback_early_stopping}.}
}
\value{
This function returns the history. The model
 is updated in situ.
}
\description{
This wrapper automatically compiles and fits a model on
 the provided training data. Loss is selected automatically
 for each output using \code{type2loss}.
}
