% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-constraint.R
\name{constraint_all_ones}
\alias{constraint_all_ones}
\title{Constrain All Weights to One}
\usage{
constraint_all_ones(w)
}
\arguments{
\item{w}{A weights matrix.}
}
\description{
This weights constraint function forces all weights to one.
}
