% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-constraint.R
\name{constraint_all_zeros}
\alias{constraint_all_zeros}
\title{Constrain All Weights to Zero}
\usage{
constraint_all_zeros(w)
}
\arguments{
\item{w}{A weights matrix.}
}
\description{
This weights constraint function forces all weights to zero.
}
