% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-constraint.R
\name{constraint_rows_to_unit_sum}
\alias{constraint_rows_to_unit_sum}
\title{Constrain Rows to Unit Sum}
\usage{
constraint_rows_to_unit_sum(w)
}
\arguments{
\item{w}{A weights matrix.}
}
\description{
This weights constraint function ensures that each row
 sums to 1. The interpretation here is that each edge from
 a node in the previous layer emits a part of the whole.
}
