% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-constraint.R
\name{constraint_runif}
\alias{constraint_runif}
\title{Constrain All Weights to Randomize}
\usage{
constraint_runif(w)
}
\arguments{
\item{w}{A weights matrix.}
}
\description{
This weights constraint function forces all weights to randomize each cycle.
}
