% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-get.R
\name{get_incoming_layer_name}
\alias{get_incoming_layer_name}
\title{Get Incoming Layer Name}
\usage{
get_incoming_layer_name(object)
}
\arguments{
\item{object}{A keras layer.}
}
\value{
A string. The layer name.
}
\description{
This function returns the name of the incoming layer.
}
