% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-AM.R
\name{get_layer_gradient}
\alias{get_layer_gradient}
\title{Get the Gradient Associated with a Layer}
\usage{
get_layer_gradient(model, loss)
}
\arguments{
\item{model}{A keras model.}

\item{loss}{The output from \code{\link{get_layer_loss}}.}
}
\value{
A gradient.
}
\description{
Get the Gradient Associated with a Layer
}
