% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-AM.R
\name{get_layer_loss}
\alias{get_layer_loss}
\title{Get the Loss Associated with a Layer}
\usage{
get_layer_loss(model, layer, filter = NA)
}
\arguments{
\item{model}{A keras model.}

\item{layer}{The name of the layer.}

\item{filter}{The filter to select. Skip with \code{filter = NA}.}
}
\value{
A scalar loss.
}
\description{
Get the Loss Associated with a Layer
}
