% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{get_layer_names}
\alias{get_layer_names}
\title{Get Layer Names}
\usage{
get_layer_names(model)
}
\arguments{
\item{model}{A keras model.}
}
\value{
A character vector. The layer names.
}
\description{
This function returns the names of all layers in a model.
}
