% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-get.R
\name{get_layer_output}
\alias{get_layer_output}
\title{Get Layer Output}
\usage{
get_layer_output(model, data, layer)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{data}{A matrix or list of matrices. The input data.}

\item{layer}{The layer name or index.}
}
\value{
An R array.
}
\description{
This function returns the output of a layer for any input data.
}
