% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-get.R
\name{get_layer_weights}
\alias{get_layer_weights}
\title{Get Layer Weights}
\usage{
get_layer_weights(model, layer)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{layer}{The layer name or index.}
}
\value{
An R array.
}
\description{
This function returns the weights of a layer.
}
