% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{layer2index}
\alias{layer2index}
\title{Find Index for Layer Name}
\usage{
layer2index(model, layer)
}
\arguments{
\item{model}{A keras model.}

\item{layer}{A string or integer. The requested layer.}
}
\value{
An integer. The requested layer's index.
}
\description{
This function returns the numeric index that corresponds
 to a requested layer. If the request is a string, this function
 looks up the index. If the request is an integer, this function
 returns that integer.
}
