% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-learnable.R
\name{layer_kernel_dot}
\alias{layer_kernel_dot}
\title{Perform Kernel Dot Product}
\usage{
layer_kernel_dot(layer, kernel, name = NULL)
}
\arguments{
\item{layer}{The incoming layer.}

\item{kernel}{A layer that represents a kernel.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
The result of \code{k_dot}.
}
\description{
This function calls \code{k_dot} as a "lambda" layer.
}
\examples{
library(keras)
library(caress)
set.seed(1)
index <- sample(1:nrow(iris))
x <- as.matrix(iris[index,1:4])
y <- to_categorical(as.numeric(iris[index,5])-1)
k_clear_session()
input <- from_input(x)
weight <- input \%>\%
  layer_learnable_array(4) \%>\%
  layer_reshape(c(4,1))
target <- layer_kernel_dot(input, weight) \%>\%
  layer_flatten() \%>\%
  to_output(y)
m <- prepare(input, target)
build(m, x, y, batch_size = 4, epochs = 100)
}
