% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-learnable.R
\name{layer_learnable_array}
\alias{layer_learnable_array}
\title{Initialize Learnable Layer}
\usage{
layer_learnable_array(input, array_dim, name = NULL)
}
\arguments{
\item{input}{The incoming layer.}

\item{array_dim}{The dimensions of the learnable kernel.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
A layer of learnable weights.
}
\description{
This function creates a layer of learnable weights. It requires connection
 to an input layer, but does not use the input. Instead, the values of the
 layer are the manifestation of learnable weights.
}
\examples{
library(keras)
library(caress)
set.seed(1)
index <- sample(1:nrow(iris))
x <- as.matrix(iris[index,1:4])
y <- to_categorical(as.numeric(iris[index,5])-1)
k_clear_session()
input <- from_input(x)
weight <- input \%>\%
  layer_learnable_array(4) \%>\%
  layer_reshape(c(4,1))
target <- layer_kernel_dot(input, weight) \%>\%
  layer_flatten() \%>\%
  to_output(y)
m <- prepare(input, target)
build(m, x, y, batch_size = 4, epochs = 100)
}
