% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-orthogonal.R
\name{layer_orthogonal_to}
\alias{layer_orthogonal_to}
\title{Create Orthogonal Layer}
\usage{
layer_orthogonal_to(target, reference, name = NULL)
}
\arguments{
\item{target}{The incoming layer with predictive power.}

\item{reference}{The incoming layer you do not want to predict.}

\item{name}{A string. The name of the layers.}
}
\value{
A layer that equals the "reference" when the "target" is
 orthogonal to the "reference".
}
\description{
This function makes the "target" layer orthogonal to the "reference"
 layer so that it cannot predict the "reference". In other words, the
 "target" layer will equal the "reference" layer when the "target"
 is orthogonal to the "reference".
}
