% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-pairwise.R
\name{layer_pairwise_residual}
\alias{layer_pairwise_residual}
\title{Create Pairwise Layer}
\usage{
layer_pairwise_residual(object, name = NULL)
}
\arguments{
\item{object}{The incoming layer.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
A layer of the pairwise residuals.
}
\description{
This function computes all pairwise residuals in the final dimension
 of a tensor. For example, if the tensor has the dimensions (2, 3, 4),
 this function will return a tensor with the dimensions (2, 3, 4, 4).
 The element (2, 3, i, j) would contain the value
 ((2, 3, i) - (2, 3, j))^2.
}
