% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-pairwise.R
\name{layer_pairwise_rmse}
\alias{layer_pairwise_rmse}
\title{Create Pairwise Layer}
\usage{
layer_pairwise_rmse(object, name = NULL)
}
\arguments{
\item{object}{The incoming layer.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
A layer of the root mean square error.
}
\description{
This function computes all pairwise residuals in the final dimension
 of a tensor. For example, if the tensor has the dimensions (2, 3, 4),
 this function will return a tensor with the dimensions (2, 3, 1).
 The element (2, 3, 1) would contain the RMSE from the input
 SQRT(MEAN(((2, 3, i) - (2, 3, j))^2)).
}
