% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-pseudo-embed.R
\name{layer_pseudo_embed}
\alias{layer_pseudo_embed}
\title{Create an Embedding Matrix}
\usage{
layer_pseudo_embed(object, embed_dim, random_embedding = FALSE, name = NULL)
}
\arguments{
\item{object}{A \code{keras} model.}

\item{embed_dim}{An integer. The size of the final embedding matrix will
equal the input dimension times the embedding dimension.}

\item{random_embedding}{A boolean. Toggles whether to freeze the embedding
matrix with random values. Otherwise, the embedding matrix is trainable.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
A layer that acts like an embedding matrix.
}
\description{
This function creates an embedding matrix in the form of a layer.
 This layer can be multiplied with other layers as if it were itself
 a weights matrix. This is because all samples within the embedding layer
 will always have the same activation.
}
