% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-DeepTRIAGE.R
\name{layer_to_dense_DeepTRIAGE}
\alias{layer_to_dense_DeepTRIAGE}
\title{Apply a DeepTRIAGE Layer}
\usage{
layer_to_dense_DeepTRIAGE(object, result_dim, embed_dim = result_dim * 4,
  random_embedding = FALSE, hidden_dim = 32,
  hidden_activation = "tanh", hidden_dropout = 0.5, name = NULL)
}
\arguments{
\item{object}{A \code{keras} model.}

\item{result_dim}{An integer. The size of the final layer.}

\item{embed_dim}{An integer. The size of the final embedding matrix will
equal the input dimension times the embedding dimension.}

\item{random_embedding}{A boolean. Toggles whether to freeze the embedding
matrix with random values. Otherwise, the embedding matrix is trainable.}

\item{hidden_dim}{An integer. The size of the hidden layers.}

\item{hidden_activation}{A string. The activation for the hidden layers.}

\item{hidden_dropout}{A numeric. The dropout for the hidden layers.}

\item{name}{A string. The prefix label for all layers.}
}
\description{
This function applies a variant of the DeepTRIAGE attention
 mechanism to the incoming layer (see DOI:10.1101/533406).
 This implementation differs slightly from the publication
 in that all layers have the same activation function and
 the random embedding weights are optionally learnable.
}
