% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-stereo.R
\name{layer_to_dense_stereo}
\alias{layer_to_dense_stereo}
\title{Split Model into Two Parallel Layers}
\usage{
layer_to_dense_stereo(object, units, activation = NULL, name = NULL,
  ...)
}
\arguments{
\item{object}{A \code{keras} model.}

\item{units, activation, ...}{Arguments passed to \code{keras::layer_dense}.}

\item{name}{A string. The prefix label for all layers.}
}
\value{
A list of two layers: Channel 1 and Channel 2.
}
\description{
This function splits the input into two parallel layers.
 Each layer has the same architecture, but is influenced
 by different weights.
}
