% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-stereo.R
\name{layer_to_dense_stereo_and_add}
\alias{layer_to_dense_stereo_and_add}
\title{Split Model into Two Parallel Layers}
\usage{
layer_to_dense_stereo_and_add(object, units, activation = NULL,
  name = NULL, ...)
}
\arguments{
\item{object}{A \code{keras} model.}

\item{units}{Arguments passed to \code{keras::layer_dense}.}

\item{activation}{Arguments passed to \code{keras::layer_dense}.}

\item{name}{A string. The prefix label for all layers.}

\item{...}{Arguments passed to \code{keras::layer_dense}.}
}
\value{
A list of three layers: Channel 1, Channel 2, and
 Channel 1 & Channel 2.
}
\description{
This function splits the input into three parallel layers.
 The first two layers have the same architecture, but are influenced
 by different weights. The third layer is the sum of
 the other two layers.
}
