% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-get.R
\name{model_decode}
\alias{model_decode}
\title{Decode Output of Any Layer}
\usage{
model_decode(model, data, input_at, output_at)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{data}{A matrix or list of matrices. The input data.
See \code{\link{get_layer_output}}.}

\item{input_at}{The layer name or index. From where the
input data has come.}

\item{output_at}{The layer name or index. From where the
output data should come.}
}
\value{
An R array.
}
\description{
This function returns the output of any one layer based on
 the input of any other layer. Use this function to decode
 the latent space of an autoencoder.
}
