% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-set.R
\name{model_mirror}
\alias{model_mirror}
\title{Copy Weights from Another Model}
\usage{
model_mirror(model, reference, freeze)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{reference}{Another \code{keras} model. The model from which
to copy the weights.}

\item{freeze}{A logical. Toggles whether or not to freeze the weights.
If frozen, the weights will not update during re-training.}
}
\value{
Null. This function updates the model in place.
}
\description{
This function uses a pre-trained "reference" model to automatically
 set the weights for all layers that have the same name.
}
