% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-build.R
\name{prepare}
\alias{prepare}
\title{Prepare a Keras Model}
\usage{
prepare(input, output)
}
\arguments{
\item{input}{A layer or list of layers. The input layer(s).}

\item{output}{A layer or list of layers. The output layer(s).}
}
\value{
A keras model.
}
\description{
This wrapper joins the input and output layer(s) using
 \code{keras::keras_model}, then plots the model graph and
 prints the summary.
}
