% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-set.R
\name{set_layer_weights}
\alias{set_layer_weights}
\title{Set Layer Weights}
\usage{
set_layer_weights(model, layer, weights, freeze)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{layer}{The layer name or index.}

\item{weights}{A matrix or list of matrices. The weights.
For example, the output of \code{\link{get_layer_weights}}.}

\item{freeze}{A logical. Toggles whether or not to freeze the weights.
If frozen, the weights will not update during re-training.}
}
\value{
Null. This function updates the model in place.
}
\description{
This function sets the weights of a layer.
}
