% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-preprocess.R
\name{to_categorical_invert}
\alias{to_categorical_invert}
\title{Make Categorical Mask}
\usage{
to_categorical_invert(y, mask = unique(y))
}
\arguments{
\item{y}{The outcome. An integer vector.}

\item{mask}{The values to mask. If missing, all values are masked.}
}
\value{
A matrix like \code{keras::to_categorical}.
}
\description{
This function one-hot encodes an integer vector as a matrix,
 except that it masks certain outcomes. For example, consider
 the three outcomes {0, 1, 2} where 2 is masked. The value 0
 is coded as {1, 0, 0}. The value 1 is coded as {0, 1, 0}.
 The value 2 is coded as {.5, .5, 0}. This blinds the
 classifier to the value 2.
}
