% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{to_metric}
\alias{to_metric}
\title{Get Metric for Output}
\usage{
to_metric(y)
}
\arguments{
\item{y}{A matrix or vector. The output data.}
}
\value{
A metric.
}
\description{
This function guesses the metric to use based on the
 type of outcome provided to \code{y}. It chooses from
 "accuracy" or "mean absolute percent error".
}
