% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{type2loss}
\alias{type2loss}
\title{Get Loss for Type}
\usage{
type2loss(type)
}
\arguments{
\item{type}{Choose from "binary", "multiclass",
"multilabel", "multivariate", or "univariate".}
}
\value{
A loss function.
}
\description{
This function guesses the loss function based on the
 type of outcome returned from \code{type_of_y}. It
 chooses from cross-entropy or mean squared error.
}
