% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{type_of_y}
\alias{type_of_y}
\title{Get Type for Y}
\usage{
type_of_y(y)
}
\arguments{
\item{y}{A matrix or vector. The output data.}
}
\value{
A string.
}
\description{
This function guesses the "type" of outcome based on the
 object provided. It guesses "one-hot-encoded", "multi-label",
 or "continuous". Used by \code{\link{to_loss}} and
 \code{\link{to_output}}.
}
