% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-build.R
\name{build}
\alias{build}
\title{Train a Keras Model}
\usage{
build(
  model,
  x_train,
  y_train,
  lr = 0.001,
  loss_weights = NULL,
  epochs = 30,
  batch_size = 128,
  early_stopping_monitor = "val_loss",
  early_stopping_patience = 5,
  reduce_lr_monitor = "val_loss",
  reduce_lr_patience = epochs,
  validation_split = 0.2
)
}
\arguments{
\item{model}{A keras model.}

\item{x_train, y_train}{A numeric or list of numerics.
Arguments passed to \code{keras::fit}.}

\item{lr, loss_weights}{Arguments passed to \code{keras::compile}.}

\item{epochs, batch_size, validation_split}{Arguments passed to \code{keras::fit}.}

\item{early_stopping_monitor, early_stopping_patience}{Arguments passed to \code{keras::callback_early_stopping}.}

\item{reduce_lr_monitor, reduce_lr_patience}{Arguments passed to \code{keras::callback_reduce_lr_on_plateau}.}
}
\value{
This function returns the history. The model
 is updated in situ.
}
\description{
This wrapper automatically compiles and fits a model on
 the provided training data. Loss is selected automatically
 for each output using \code{type2loss}.
}
