% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-in-out.R
\name{from_input}
\alias{from_input}
\title{Create Input Layer}
\usage{
from_input(x, name = NULL)
}
\arguments{
\item{x}{A matrix or list of matrices. The input data.}

\item{name}{A string or character vector. The name(s) of the layer.}
}
\value{
A layer object (or list of layer objects).
}
\description{
This function automatically creates an input layer for
 a provided matrix object.
}
