% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-sample_random.R
\name{sample_random}
\alias{sample_random}
\title{Split the Training and Test Set}
\usage{
sample_random(x, y, split = 67, normalize = TRUE)
}
\arguments{
\item{x}{The input data or a list of input data.}

\item{y}{The output data or a list of output data.}

\item{split}{An integer. The training-test set split as a percentage.}

\item{normalize}{A logical. Toggles whether to normalize each
feature by subtracting the training set mean and dividing
by the training set standard deviation.}
}
\value{
A nested list with two slots "train" and "test", each of which
 contain another list with two more slots "x" and "y". These provide a
 list of the processed inputs and outputs, respectively.
 For example, access the training input with $train$x.
 You can feed these lists to \code{build}.
}
\description{
This function splits a list of inputs and a list of outputs
 by random sampling. All inputs and outputs get split the same.
}
