% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-in-out.R
\name{to_output}
\alias{to_output}
\title{Create Output Layer}
\usage{
to_output(model, y, name = NULL, append_name = FALSE)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{y}{A matrix or vector. The output data.}

\item{name}{A string. The name of the layer.}

\item{append_name}{A boolean. Toggles whether to give each
name a unique appendix. Used for backend only.}
}
\value{
A layer object.
}
\description{
This function automatically creates an output layer for
 a provided matrix object.
}
