% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-layers-in-out.R
\name{to_output}
\alias{to_output}
\title{Create Output Layer}
\usage{
to_output(model, y, name = NULL)
}
\arguments{
\item{model}{A \code{keras} model.}

\item{y}{A matrix or vector. The output data.}

\item{name}{A string. The name of the layer.}
}
\value{
A layer object.
}
\description{
This function automatically creates an output layer for
 a provided matrix object.
}
