% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-backend.R
\name{type_of_y}
\alias{type_of_y}
\title{Get Type for Y}
\usage{
type_of_y(y)
}
\arguments{
\item{y}{The outcome. A vector or matrix.}
}
\value{
A string.
}
\description{
This function guesses the "type" of outcome based on the
 object provided. It guesses "binary", "multiclass",
 "multilabel", "multivariate", or "univariate". Used by
 \code{type2loss} to choose a loss function.
}
